+++ Changelog.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [Unerleased]
 ### Fixed
+ - Sketchpad: issue in Chrome with background image being displayed as text
  - FOR-2239: Sketchpad -> Fixed issues with background being shifted in datagrid
 
 ## 3.20.10